+++ CHANGELOG.rst
@@ -3,6 +3,8 @@
 
     * Error message in ``Signal.chpoints`` setter.
     * Typing of ``Signal.y``.
+    * Validate in ``Signal`` that ``y``, ``t`` and ``marks`` value is 1D
+      (to prevent scalar values from wrongly called ``numpy`` functions).
 
 0.0.1 (2023-01-12)
     First release.